Extract some methods to module
Fix RuboCop `Metrics/ModuleLength` cop
