+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2018-12-06 14:24:14 +0300 using RuboCop version 0.61.1.
+# on 2019-01-25 13:34:12 +0300 using RuboCop version 0.63.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -37,15 +37,10 @@ Metrics/ClassLength:
 Metrics/CyclomaticComplexity:
   Max: 42
 
-# Offense count: 192
+# Offense count: 193
 # Configuration parameters: CountComments, ExcludedMethods.
 Metrics/MethodLength:
   Max: 83
-
-# Offense count: 1
-# Configuration parameters: CountComments.
-Metrics/ModuleLength:
-  Max: 114
 
 # Offense count: 2
 # Configuration parameters: CountKeywordArgs.
@@ -60,7 +55,7 @@ Metrics/PerceivedComplexity:
 Style/Documentation:
   Enabled: false
 
-# Offense count: 1749
+# Offense count: 1763
 # Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
 # URISchemes: http, https
 Metrics/LineLength:
