+++ lib/ooxml_parser/common_parser/common_data/ooxml_document_object/ooxml_document_object_helper.rb
@@ -105,25 +105,6 @@ module OoxmlDocumentObjectHelper
     symbol
   end
 
-  def option_enabled?(node, attribute_name = 'val')
-    return true if node.attributes.empty?
-    return true if node.to_s == '1'
-    return false if node.to_s == '0'
-    return false if node.attribute(attribute_name).nil?
-
-    status = node.attribute(attribute_name).value
-    !%w[false off 0].include?(status)
-  end
-
-  def attribute_enabled?(node, attribute_name = 'val')
-    return true if node.to_s == '1'
-    return false if node.to_s == '0'
-    return false if node.attribute(attribute_name).nil?
-
-    status = node.attribute(attribute_name).value
-    %w[true on 1].include?(status)
-  end
-
   def root_object
     tree_object = self
     tree_object = tree_object.parent until tree_object.parent.nil?
